pyproject.toml
[tool.poetry]
name = "etf-assistant"
version = "0.1.0"
description = ""
authors = ["Simone Eandi"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
chromadb = "^0.4.22"
langchain = "^0.1.5"
python-dotenv = "^1.0.1"
openai = "^1.11.0"
jq = "^1.6.0"
tiktoken = "^0.5.2"
langfuse = "^2.9.0"
loguru = "^0.7.2"
langchain-openai = "^0.0.5"
faiss-cpu = "^1.7.4"
pdfplumber = "^0.10.4"
unstructured = {extras = ["all-docs"], version = "^0.12.4"}
pymupdf = "^1.23.25"
html2text = "^2024.2.25"
etfpy = "^1.0.16"
rapidocr-onnxruntime = "^1.3.15"
minio = "^7.2.5"
streamlit = "^1.32.2"
streamlit-pdf-viewer = "^0.0.8"
levenshtein = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
